refactor: initializeWorkspace command and related utilities
