+++ src/command/command/index.mts
@@ -1,5 +1,11 @@
 import collectAdmonitions from "./collectAdmonitions/index.mjs";
 import createNote from "./createNote/index.mjs";
+import initializeWorkspace from "./initializeWorkspace/index.mjs";
 import resetGlobalFleximarkCss from "./resetGlobalFleximarkCss/index.mjs";
 
-export { collectAdmonitions, createNote, resetGlobalFleximarkCss };
+export {
+  collectAdmonitions,
+  createNote,
+  initializeWorkspace,
+  resetGlobalFleximarkCss,
+};
