+++ test/command/lib/settings/index.test.mts
