+++ test/extension.test.mts
@@ -3,8 +3,8 @@ import * as assert from "assert";
 // as well as import your extension to test it
 import * as vscode from "vscode";
 
-import * as commands_genSettingsJson_index from "./command/getSettingsJson/index.test.mjs";
 import * as commands_css_index from "./command/lib/css/index.test.mjs";
+import * as commands_genSettingsJson_index from "./command/lib/settings/index.test.mjs";
 import * as commands_utils_getBlockLineAndOffset from "./command/utils/getBlockLineAndOffset.test.mjs";
 import * as completion_lib_checkCurrentLineLangMode from "./completion/lib/checkCurrentLineLangMode.test.mjs";
